refactor: split packet into multiple files
